+++ packages/exchange-consumer/src/consumers/index.ts
@@ -1,1 +1,2 @@
 export * from './TransactionConsumer'
+export * from './InsuranceFundConsumer'