+++ packages/exchange-consumer/src/types.ts
@@ -1,3 +1,4 @@
 import { MarketStatusMap as GrpcMarketStatus } from '@injectivelabs/chain-api/injective/exchange/v1beta1/exchange_pb'
+import { InsuranceFund as GrpcInsuranceFund } from '@injectivelabs/exchange-api/injective_insurance_rpc_pb'
 
-export { GrpcMarketStatus }
+export { GrpcMarketStatus, GrpcInsuranceFund }